Move add data and macro settings to common
