+++ python/pseudoBinaryPhaseDiagramGUI.py
@@ -3,7 +3,6 @@ import PySimpleGUI as sg
 import matplotlib.pyplot as plt
 import os
 import sys
-import shutil
 import copy
 import thermoToolsGUI
 
@@ -227,10 +226,10 @@ class CalculationWindow:
             self.sgw.Element('Add Data').Update(disabled = False)
         elif event =='Add Data':
             self.calculation.makeBackup()
-            addDataWindow = AddDataWindow(self)
-            self.children.append(addDataWindow)
+            addDataWindow = thermoToolsGUI.AddDataWindow(self)
+            self.children.append(addDataWindow,windowList)
         elif event =='Macro Settings':
-            macroSettingsWindow = MacroSettingsWindow(self)
+            macroSettingsWindow = thermoToolsGUI.MacroSettingsWindow(self,windowList)
             self.children.append(macroSettingsWindow)
     def makeLayout(self):
         tempLayout = [sg.Column([[sg.Text('Temperature')],[sg.Input(key='-temperature-',size=(thermoToolsGUI.inputSize,1))],
@@ -570,151 +569,6 @@ class SettingsWindow:
                 pass
             self.parent.calculation.makePlot()
             self.close()
-
-class AddDataWindow:
-    def __init__(self,parent):
-        self.parent = parent
-        windowList.append(self)
-        file_list_column = [
-            [
-                sg.Text("Experimental Data Folder"),
-                sg.In(size=(25, 1), enable_events=True, key="-FOLDER-"),
-                sg.FolderBrowse(),
-            ],
-            [
-                sg.Listbox(
-                    values=[], enable_events=True, size=(40, 20), key="-FILE LIST-"
-                )
-            ],
-        ]
-        self.folder = os.getcwd()
-        try:
-            file_list = os.listdir(self.folder)
-        except:
-            file_list = []
-        buttonLayout = [sg.Button('Add Data'), sg.Button('Exit')]
-        addDataLayout = [file_list_column,buttonLayout]
-        self.sgw = sg.Window('Experimental data selection', addDataLayout, location = [0,0], finalize=True)
-        fnames = [
-            f
-            for f in file_list
-            if os.path.isfile(os.path.join(self.folder, f))
-            and f.lower().endswith((".csv"))
-        ]
-        fnames = sorted(fnames, key=str.lower)
-        self.sgw["-FILE LIST-"].update(fnames)
-        self.children = []
-        self.filename = ''
-    def close(self):
-        for child in self.children:
-            child.close()
-        self.sgw.close()
-        if self in windowList:
-            windowList.remove(self)
-    def read(self):
-        event, values = self.sgw.read(timeout=thermoToolsGUI.timeout)
-        if event == sg.WIN_CLOSED or event == 'Exit':
-            self.close()
-        elif event == "-FOLDER-":
-            self.filename = ''
-            self.folder = values["-FOLDER-"]
-            try:
-                file_list = os.listdir(self.folder)
-            except:
-                file_list = []
-
-            fnames = [
-                f
-                for f in file_list
-                if os.path.isfile(os.path.join(self.folder, f))
-                and f.lower().endswith((".csv"))
-            ]
-            fnames = sorted(fnames, key=str.lower)
-            self.sgw["-FILE LIST-"].update(fnames)
-        elif event == "-FILE LIST-":  # A file was chosen from the listbox
-            self.filename = values["-FILE LIST-"][0]
-        elif event == 'Add Data':
-            if not self.filename:
-                return
-            datafile = os.path.join(self.folder, self.filename)
-            expName = self.filename.split('.',1)[0]
-            self.parent.calculation.addData(datafile,expName)
-            self.parent.macro.append(f'macroPD.addData("{datafile}","{expName}")')
-
-class MacroSettingsWindow:
-    def __init__(self,parent):
-        self.parent = parent
-        windowList.append(self)
-        file_list_column = [
-            [
-                sg.Text("Macro Folder"),
-                sg.In(size=(25, 1), enable_events=True, key="-FOLDER-"),
-                sg.FolderBrowse(),
-            ],
-            [
-                sg.Listbox(
-                    values=[], enable_events=True, size=(40, 20), key="-FILE LIST-"
-                )
-            ],
-        ]
-        self.folder = os.getcwd() + '/python'
-        try:
-            file_list = os.listdir(self.folder)
-        except:
-            file_list = []
-        buttonLayout = [sg.Button('Select Macro', size = thermoToolsGUI.buttonSize)]
-        inputNameLayout = [[sg.Text('Macro File Save Name:')],[sg.Input(key='-macroSaveName-',size=(thermoToolsGUI.inputSize,1)),sg.Text('.py')],[sg.Button('Set Save Name', size = thermoToolsGUI.buttonSize), sg.Button('Exit', size = thermoToolsGUI.buttonSize)]]
-        addDataLayout = [file_list_column,buttonLayout,inputNameLayout]
-        self.sgw = sg.Window('Macro file', addDataLayout, location = [0,0], finalize=True)
-        fnames = [
-            f
-            for f in file_list
-            if os.path.isfile(os.path.join(self.folder, f))
-            and f.lower().endswith((".py"))
-        ]
-        fnames = sorted(fnames, key=str.lower)
-        self.sgw["-FILE LIST-"].update(fnames)
-        self.children = []
-        self.filename = ''
-    def close(self):
-        for child in self.children:
-            child.close()
-        self.sgw.close()
-        if self in windowList:
-            windowList.remove(self)
-    def read(self):
-        event, values = self.sgw.read(timeout=thermoToolsGUI.timeout)
-        if event == sg.WIN_CLOSED or event == 'Exit':
-            self.close()
-        elif event == "-FOLDER-":
-            self.filename = ''
-            self.folder = values["-FOLDER-"]
-            try:
-                file_list = os.listdir(self.folder)
-            except:
-                file_list = []
-
-            fnames = [
-                f
-                for f in file_list
-                if os.path.isfile(os.path.join(self.folder, f))
-                and f.lower().endswith((".py"))
-            ]
-            fnames = sorted(fnames, key=str.lower)
-            self.sgw["-FILE LIST-"].update(fnames)
-        elif event == "-FILE LIST-":  # A file was chosen from the listbox
-            self.filename = values["-FILE LIST-"][0]
-        elif event == 'Select Macro':
-            if not self.filename:
-                return
-            datafile = os.path.join(self.folder, self.filename)
-            # I don't want to mess with import logic, so rename/overwrite file in one spot instead
-            shutil.copy(datafile,os.getcwd() + '/python/' + 'macroPhaseDiagram.py')
-            self.close()
-        elif event == 'Set Save Name':
-            if not values["-macroSaveName-"]:
-                return
-            self.parent.macroSaveName = values["-macroSaveName-"] + '.py'
 
 if not(os.path.isfile('bin/InputScriptMode')):
     errorLayout = [[sg.Text('No Thermochimica executable available.')],
