+++ python/thermoToolsGUI.py
@@ -2,6 +2,7 @@ import PySimpleGUI as sg
 import os
 import math
 import thermoTools
+import shutil
 
 timeout = 50
 inputSize = 20
@@ -116,3 +117,150 @@ class DataWindow:
                 return
             calcWindow = self.calc(self,datafile,nElements,elements,True)
             self.children.append(calcWindow)
+
+class AddDataWindow:
+    def __init__(self,parent,windowList):
+        self.parent = parent
+        self.windowList = windowList
+        windowList.append(self)
+        file_list_column = [
+            [
+                sg.Text("Experimental Data Folder"),
+                sg.In(size=(25, 1), enable_events=True, key="-FOLDER-"),
+                sg.FolderBrowse(),
+            ],
+            [
+                sg.Listbox(
+                    values=[], enable_events=True, size=(40, 20), key="-FILE LIST-"
+                )
+            ],
+        ]
+        self.folder = os.getcwd()
+        try:
+            file_list = os.listdir(self.folder)
+        except:
+            file_list = []
+        buttonLayout = [sg.Button('Add Data'), sg.Button('Exit')]
+        addDataLayout = [file_list_column,buttonLayout]
+        self.sgw = sg.Window('Experimental data selection', addDataLayout, location = [0,0], finalize=True)
+        fnames = [
+            f
+            for f in file_list
+            if os.path.isfile(os.path.join(self.folder, f))
+            and f.lower().endswith((".csv"))
+        ]
+        fnames = sorted(fnames, key=str.lower)
+        self.sgw["-FILE LIST-"].update(fnames)
+        self.children = []
+        self.filename = ''
+    def close(self):
+        for child in self.children:
+            child.close()
+        self.sgw.close()
+        if self in self.windowList:
+            self.windowList.remove(self)
+    def read(self):
+        event, values = self.sgw.read(timeout=timeout)
+        if event == sg.WIN_CLOSED or event == 'Exit':
+            self.close()
+        elif event == "-FOLDER-":
+            self.filename = ''
+            self.folder = values["-FOLDER-"]
+            try:
+                file_list = os.listdir(self.folder)
+            except:
+                file_list = []
+
+            fnames = [
+                f
+                for f in file_list
+                if os.path.isfile(os.path.join(self.folder, f))
+                and f.lower().endswith((".csv"))
+            ]
+            fnames = sorted(fnames, key=str.lower)
+            self.sgw["-FILE LIST-"].update(fnames)
+        elif event == "-FILE LIST-":  # A file was chosen from the listbox
+            self.filename = values["-FILE LIST-"][0]
+        elif event == 'Add Data':
+            if not self.filename:
+                return
+            datafile = os.path.join(self.folder, self.filename)
+            expName = self.filename.split('.',1)[0]
+            self.parent.calculation.addData(datafile,expName)
+            self.parent.macro.append(f'macroPD.addData("{datafile}","{expName}")')
+
+class MacroSettingsWindow:
+    def __init__(self,parent,windowList):
+        self.parent = parent
+        self.windowList = windowList
+        windowList.append(self)
+        file_list_column = [
+            [
+                sg.Text("Macro Folder"),
+                sg.In(size=(25, 1), enable_events=True, key="-FOLDER-"),
+                sg.FolderBrowse(),
+            ],
+            [
+                sg.Listbox(
+                    values=[], enable_events=True, size=(40, 20), key="-FILE LIST-"
+                )
+            ],
+        ]
+        self.folder = os.getcwd() + '/python'
+        try:
+            file_list = os.listdir(self.folder)
+        except:
+            file_list = []
+        buttonLayout = [sg.Button('Select Macro', size = buttonSize)]
+        inputNameLayout = [[sg.Text('Macro File Save Name:')],[sg.Input(key='-macroSaveName-',size=(inputSize,1)),sg.Text('.py')],[sg.Button('Set Save Name', size = buttonSize), sg.Button('Exit', size = buttonSize)]]
+        addDataLayout = [file_list_column,buttonLayout,inputNameLayout]
+        self.sgw = sg.Window('Macro file', addDataLayout, location = [0,0], finalize=True)
+        fnames = [
+            f
+            for f in file_list
+            if os.path.isfile(os.path.join(self.folder, f))
+            and f.lower().endswith((".py"))
+        ]
+        fnames = sorted(fnames, key=str.lower)
+        self.sgw["-FILE LIST-"].update(fnames)
+        self.children = []
+        self.filename = ''
+    def close(self):
+        for child in self.children:
+            child.close()
+        self.sgw.close()
+        if self in self.windowList:
+            self.windowList.remove(self)
+    def read(self):
+        event, values = self.sgw.read(timeout=timeout)
+        if event == sg.WIN_CLOSED or event == 'Exit':
+            self.close()
+        elif event == "-FOLDER-":
+            self.filename = ''
+            self.folder = values["-FOLDER-"]
+            try:
+                file_list = os.listdir(self.folder)
+            except:
+                file_list = []
+
+            fnames = [
+                f
+                for f in file_list
+                if os.path.isfile(os.path.join(self.folder, f))
+                and f.lower().endswith((".py"))
+            ]
+            fnames = sorted(fnames, key=str.lower)
+            self.sgw["-FILE LIST-"].update(fnames)
+        elif event == "-FILE LIST-":  # A file was chosen from the listbox
+            self.filename = values["-FILE LIST-"][0]
+        elif event == 'Select Macro':
+            if not self.filename:
+                return
+            datafile = os.path.join(self.folder, self.filename)
+            # I don't want to mess with import logic, so rename/overwrite file in one spot instead
+            shutil.copy(datafile,os.getcwd() + '/python/' + 'macroPhaseDiagram.py')
+            self.close()
+        elif event == 'Set Save Name':
+            if not values["-macroSaveName-"]:
+                return
+            self.parent.macroSaveName = values["-macroSaveName-"] + '.py'
